requirements.txt
discord.py==1.4.1
discord.py[voice]==1.4.1
gTTS==2.1.1
dictdiffer~=0.8.1